// src/app/cc-color-picker/model/colorPaletteType.ts
export enum ColorPaletteType {
  None = 0,
  Red = 1,
  Pink = 2,
  Purple = 3,
  DeepPurple = 4,
  Indigo = 5,
  Blue = 6,
  LightBlue = 7,
  Cyan = 8,
  Teal = 9,
  Green = 10,
  LightGreen = 11,
  Lime = 12,
  Yellow = 13,
  Amber = 14,
  Orange = 15,
  DeepOrange = 16,
  Brown = 17,
  Gray = 18,
  BlueGray,
}